refactor: rename ps2 keyboard driver to pc-keyboard
