+++ Cargo.toml
@@ -217,9 +217,13 @@ virtio-vsock = ["virtio"]
 vga = []
 
 ## Enables the PS/2 keyboard driver.
-##
-## This is only useful on PCs (x86-64).
-keyboard = []   
+## 
+## This feature initializes the PS/2 keyboard controller and installs a keyboard interrupt handler. 
+## It also provides a system call to receive the last scancode from the internal keyboard buffer. 
+## Note that this is not a complete keyboard driver and not needed for general keyboard support.
+## It allows receiving scancodes from the PS/2 keyboard that can be used to port applications.
+## This is only useful on PCs (x86-64). 
+pc-keyboard = []   
 
 #! ### Performance Features
