+++ src/arch/x86_64/kernel/interrupts.rs
@@ -158,9 +158,9 @@ pub(crate) fn install() {
 }
 
 pub(crate) fn install_handlers(#[allow(unused_mut)] mut handlers: InterruptHandlerMap) {
-	#[cfg(feature = "keyboard")]
+	#[cfg(feature = "pc-keyboard")]
 	{
-		use crate::arch::kernel::keyboard::get_keyboard_handler;
+		use crate::arch::kernel::pc_keyboard::get_keyboard_handler;
 		let (irq, handler) = get_keyboard_handler();
 		handlers.entry(irq).or_default().push_back(handler);
 	}
