+++ src/arch/x86_64/kernel/mod.rs
@@ -20,10 +20,10 @@ pub mod gdt;
 pub mod interrupts;
 #[cfg(feature = "kernel-stack")]
 pub mod kernel_stack;
-#[cfg(feature = "keyboard")]
-pub mod keyboard;
 #[cfg(all(not(feature = "pci"), feature = "virtio"))]
 pub mod mmio;
+#[cfg(feature = "pc-keyboard")]
+pub mod pc_keyboard;
 #[cfg(feature = "pci")]
 pub mod pci;
 pub mod pic;
