+++ src/syscalls/system.rs
@@ -7,16 +7,9 @@ pub extern "C" fn sys_getpagesize() -> i32 {
 	BasePageSize::SIZE.try_into().unwrap()
 }
 
-#[cfg(all(target_arch = "x86_64", feature = "keyboard"))]
+#[cfg(all(target_arch = "x86_64", feature = "pc-keyboard"))]
 #[hermit_macro::system]
 #[unsafe(no_mangle)]
 pub extern "C" fn sys_read_keyboard() -> u8 {
-	crate::kernel::keyboard::pop_scancode().unwrap_or(0)
-}
-
-#[cfg(not(all(target_arch = "x86_64", feature = "keyboard")))]
-#[hermit_macro::system]
-#[unsafe(no_mangle)]
-pub extern "C" fn sys_read_keyboard() -> u8 {
-	0
+	crate::kernel::pc_keyboard::pop_scancode().unwrap_or(0)
 }
